backend: wrong datatype

+++ backend/TuringMachine.Backend/TuringMachine.Backend.Server/Models/UserInterface/TransitionLines/TransitionLine.cs
@@ -3,6 +3,6 @@ namespace TuringMachine.Backend.Server.Models.UserInterface
     internal class TransitionLine
     {
         public Vector2<float>        Source { get; set; }
-        public IList<Vector2<byte>>  Steps  { get; set; }
+        public IList<Vector2<float>>  Steps  { get; set; }
     }
 }
